+++ packages/convers/plugin.js
@@ -4,10 +4,10 @@ K.Plugin({
 	templates: {
 		navSidebar: 'navSidebar_conver',
 		panelProfile: 'panelProfile_conver',
-		panelPlace: 'panelPlace_conver',
+		panelPlace: {name: 'panelPlace_conver', order: -10},
 		panelUser: 'panelUser_conver',
 		popupPlace: 'popupPlace_conver',
-		itemPlace: 'itemPlace_conver',
+		//itemPlace: 'itemPlace_conver',
 		itemUser: 'itemUser_conver'	
 	},
 	schemas: {