+++ packages/core/modules/Plugin.js
@@ -12,9 +12,12 @@ Kepler.Plugin = function(plugin) {
 		
 			if(_.isObject(plugin.templates))
 				_.each(K.templates, function(tmpls, name) {
-					console.log(tmpls)
 					if(plugin.templates[name]) {
-						K.templates[name] = _.union(tmpls, plugin.templates[name]);
+						var tt = _.union(tmpls, plugin.templates[name]);
+						
+						K.templates[name] = _.map(tt, function(v) {
+							return _.isObject(v) ? v : {name: v, order: 0};
+						});
 					}
 				});