+++ packages/share/plugin.js
@@ -2,7 +2,7 @@
 K.Plugin({
 	name: 'share',
 	templates: {
-		panelPlace: ['panelPlace_share'],
+		panelPlace: {name:'panelPlace_share',order:-5},
 		popupCursor: 'popupCursor_share',
 		//popupUser: 'popupCursor_share'
 	}