+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 To see all releases [here](https://github.com/Keplerjs/Kepler/releases)
 
-## next version
+## Next version
 - support templates custom ordering
 
 ## v1.2.5
@@ -15,11 +15,19 @@ To see all releases [here](https://github.com/Keplerjs/Kepler/releases)
 - K.Cache support auto expiration cache
 
 ## v1.2.3
+- Verification new user account by email link
 
 ## v1.2.2
+- refactoring convers plugin
 
 ## v1.2.1
+- support Meteor 1.5.1
 
 ## v1.2.0
+- support Meteor 1.4.0
 
 ## v1.1.0
+- optimize plugins pois and tracks for overpass api
+- show created places in user panel
+- show favorites place of users in panel
+- updates i18n for fr